-Wall -Werror on every C compile: 393 warning sites fixed at the root (audit N7); the warnings were hiding real bugs
Inventory: 3,166 warning lines per fresh build = 287 unique sites, plus 27
in the 57 object rules that had never been compiled with -Wall, plus 79
that only appear in specific rule contexts or on a forced full rebuild.
Fixed by correcting prototypes and declared types, checking results where
persistence or protocol writes depend on them, bounding buffers, one
statement per line, and deleting dead code -- never a pragma, never a
blanket cast. WARNFLAGS := -Wall -Werror is appended wherever HARDEN_CC is
and to the 45 object rules that had neither.

Real defects the warnings had been hiding, all fixed:
- daemon/tx_accept.c: an `unsigned` local passed as the u64* out-parameter
  of txacc_varint in three functions -- an 8-byte store into a 4-byte stack
  local, the incident #49 class.
- daemon/main.c: catch-up peer tables are [][DL_POOL_SLOT] (80 bytes) but
  four functions took char[][64]: live[] written at 64-byte stride, read at
  80 (the manual-peer check saw mangled tails), peers.good persisted as junk
  past line 1, onion entries still cut at 63 despite the widened slot.
- rpc_server.c: the reply header write() result was ignored while the body
  loop handled short writes -- a short header produced a corrupt reply.
- rpc_wallet_ops.c: watch-only descriptors over 339 chars silently
  truncated into a 340-byte field under a 512-byte line reader.
- rpc_chain.c: u64 vs uint64_t* mismatch hiding a maybe-uninitialized;
  test_rpc_chain.c's block-scope forward declaration of the mempool entry
  struct declared a NEW type, making a hook assignment type-incorrect.
- daemon/archive_verify.c: the pwrite marking an index record pruned was
  unchecked -- a failed write left the record pointing into a deleted file.
- daemon/utxo_setinfo.c: run-file names stored truncated at 64 bytes.
- tests: OP_CHECKSIG written one past two [34] arrays; a 32-byte prev-hash
  filled from an 8-byte integer; node_handshake and input_outpoint results
  never checked; test_bip30 had no prerequisite on the shim it drives.

Not covered: nasm's 22 warning lines in 6 kinds (NASMFLAGS stays without
-Werror until those are cleaned).

+++ asm/daemon/reorg.c
@@ -454,8 +454,8 @@ static int rg_hdr_at(void* vctx, long h, unsigned char hdr[80]){
         memcpy(hdr, x->c->hdr[i], 80);
         return 1;
     }
-    unsigned long long meta[3];
-    if (store_get_at(x->st, (unsigned long long)h, meta) != 1) return 0;
+    uint64_t meta[3];                       /* store_get_at's declared out type */
+    if (store_get_at(x->st, (uint64_t)h, meta) != 1) return 0;
     int fd = store_rd_fd(x->st, (unsigned)meta[2]);
     if (fd < 0) return 0;
     /* +8 skips the [len][magic] frame header -- store_read_meta's own
